Fix mixed precision for BART

+++ keras_nlp/models/bart/bart_seq_2_seq_lm.py
@@ -194,7 +194,7 @@ class BartSeq2SeqLM(GenerativeTask):
         # to vocabulary logits.
         outputs = tf.matmul(
             x,
-            backbone.token_embedding.embeddings,
+            tf.cast(backbone.token_embedding.embeddings, x.dtype),
             transpose_b=True,
         )
